// milo/ui/src/proto/infra/appengine/cr-rev/frontend/api/v1/service.pb.ts
export interface RedirectRequest {
  /**
   * query is equal to URL path of crrev main redirect logic. For example, it
   * can be "/3" (redirect to chromium/src commit with position 3) or
   * /{some_commit_hash} (redirect to some repository based on
   * some_commit_hash), etc.
   */
  readonly query: string;
}

export interface RedirectResponse {
  /**
   * git_hash is a full git commit hash of matched commit that is used for
   * redirect.
   */
  readonly gitHash: string;
  /** host is googlesource host (e.g. chromium). */
  readonly host: string;
  /** repository is Git repository (e.g. chromium/src) */
  readonly repository: string;
  /**
   * redirect_url is Gitiles URL of the commit, the same URL that user is
   * redirected to when using crrev's main redirect logic.
   */
  readonly redirectUrl: string;
}

export interface NumberingRequest {
  /** host is googlesource host (e.g. chromium). */
  readonly host: string;
  /** repository is Git repository (e.g. chromium/src) */
  readonly repository: string;
  /**
   * position_ref is name of position defined in value of git-footer git-svn-id
   * or Cr-Commit-Position (e.g. refs/heads/master,
   * svn://svn.chromium.org/chrome/trunk/src)
   */
  readonly positionRef: string;
  /**
   * position_number is sequential identifier of commit in given branch
   * (position_ref).
   */
  readonly positionNumber: string;
}

export interface NumberingResponse {
  /**
   * git_hash is a full git commit hash of matched commit that is used for
   * redirect.
   */
  readonly gitHash: string;
  /**
   * position_number matches NumberingRequest position_number. Probably useless
   * to clients, but kept for backward compatibility.
   */
  readonly positionNumber: string;
  /**
   * host matches NumberingRequest host. Probably useless to clients, but kept
   * for backward compatibility.
   */
  readonly host: string;
  /**
   * repository matches NumberingRequest repository. Probably useless to
   * clients, but kept for backward compatibility.
   */
  readonly repository: string;
}
// ...
export const RedirectRequest: MessageFns<RedirectRequest> = {
  encode(message: RedirectRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.query !== "") {
      writer.uint32(10).string(message.query);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): RedirectRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRedirectRequest() as any;
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.query = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): RedirectRequest {
    return { query: isSet(object.query) ? globalThis.String(object.query) : "" };
  },

  toJSON(message: RedirectRequest): unknown {
    const obj: any = {};
    if (message.query !== "") {
      obj.query = message.query;
    }
    return obj;
  },

  create(base?: DeepPartial<RedirectRequest>): RedirectRequest {
    return RedirectRequest.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<RedirectRequest>): RedirectRequest {
    const message = createBaseRedirectRequest() as any;
    message.query = object.query ?? "";
    return message;
  },
};
// ...
export const NumberingRequest: MessageFns<NumberingRequest> = {
  encode(message: NumberingRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.host !== "") {
      writer.uint32(10).string(message.host);
    }
    if (message.repository !== "") {
      writer.uint32(18).string(message.repository);
    }
    if (message.positionRef !== "") {
      writer.uint32(26).string(message.positionRef);
    }
    if (message.positionNumber !== "0") {
      writer.uint32(32).int64(message.positionNumber);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): NumberingRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseNumberingRequest() as any;
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.host = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.repository = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.positionRef = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.positionNumber = reader.int64().toString();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): NumberingRequest {
    return {
      host: isSet(object.host) ? globalThis.String(object.host) : "",
      repository: isSet(object.repository) ? globalThis.String(object.repository) : "",
      positionRef: isSet(object.positionRef) ? globalThis.String(object.positionRef) : "",
      positionNumber: isSet(object.positionNumber) ? globalThis.String(object.positionNumber) : "0",
    };
  },

  toJSON(message: NumberingRequest): unknown {
    const obj: any = {};
    if (message.host !== "") {
      obj.host = message.host;
    }
    if (message.repository !== "") {
      obj.repository = message.repository;
    }
    if (message.positionRef !== "") {
      obj.positionRef = message.positionRef;
    }
    if (message.positionNumber !== "0") {
      obj.positionNumber = message.positionNumber;
    }
    return obj;
  },

  create(base?: DeepPartial<NumberingRequest>): NumberingRequest {
    return NumberingRequest.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<NumberingRequest>): NumberingRequest {
    const message = createBaseNumberingRequest() as any;
    message.host = object.host ?? "";
    message.repository = object.repository ?? "";
    message.positionRef = object.positionRef ?? "";
    message.positionNumber = object.positionNumber ?? "0";
    return message;
  },
};

function createBaseNumberingResponse(): NumberingResponse {
  return { gitHash: "", positionNumber: "0", host: "", repository: "" };
}

export const NumberingResponse: MessageFns<NumberingResponse> = {
  encode(message: NumberingResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.gitHash !== "") {
      writer.uint32(10).string(message.gitHash);
    }
    if (message.positionNumber !== "0") {
      writer.uint32(16).int64(message.positionNumber);
    }
    if (message.host !== "") {
      writer.uint32(26).string(message.host);
    }
    if (message.repository !== "") {
      writer.uint32(34).string(message.repository);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): NumberingResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseNumberingResponse() as any;
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.gitHash = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.positionNumber = reader.int64().toString();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.host = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.repository = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): NumberingResponse {
    return {
      gitHash: isSet(object.git_sha) ? globalThis.String(object.git_sha) : "",
      positionNumber: isSet(object.number) ? globalThis.String(object.number) : "0",
      host: isSet(object.project) ? globalThis.String(object.project) : "",
      repository: isSet(object.repo) ? globalThis.String(object.repo) : "",
    };
  },

  toJSON(message: NumberingResponse): unknown {
    const obj: any = {};
    if (message.gitHash !== "") {
      obj.git_sha = message.gitHash;
    }
    if (message.positionNumber !== "0") {
      obj.number = message.positionNumber;
    }
    if (message.host !== "") {
      obj.project = message.host;
    }
    if (message.repository !== "") {
      obj.repo = message.repository;
    }
    return obj;
  },

  create(base?: DeepPartial<NumberingResponse>): NumberingResponse {
    return NumberingResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<NumberingResponse>): NumberingResponse {
    const message = createBaseNumberingResponse() as any;
    message.gitHash = object.gitHash ?? "";
    message.positionNumber = object.positionNumber ?? "0";
    message.host = object.host ?? "";
    message.repository = object.repository ?? "";
    return message;
  },
};
// ...
export interface Crrev {
  /**
   * Redirect implements the same logic as the main crrev redirect, but returns
   * redirect and commit information in body instead of HTTP redirect.
   */
  Redirect(request: RedirectRequest): Promise<RedirectResponse>;
  /**
   * Numbering returns commit that matches desired position of commit, based on
   * NumberingRequest parameters. Commit position is based on git-footer
   * git-svn-id or Cr-Commit-Position.
   */
  Numbering(request: NumberingRequest): Promise<NumberingResponse>;
  /**
   * Commit returns commit with desired commit hash. If there are multiple
   * commits with the same commit hash (which happens with forks and mirrors),
   * it checks priorities based on config. It is possible that priorities are
   * the same. In such case, there is no guarantee which one will be returned.
   */
  Commit(request: CommitRequest): Promise<CommitResponse>;
}

export const CrrevServiceName = "crrev.Crrev";
export class CrrevClientImpl implements Crrev {
  static readonly DEFAULT_SERVICE = CrrevServiceName;
  private readonly rpc: Rpc;
  private readonly service: string;
  constructor(rpc: Rpc, opts?: { service?: string }) {
    this.service = opts?.service || CrrevServiceName;
    this.rpc = rpc;
    this.Redirect = this.Redirect.bind(this);
    this.Numbering = this.Numbering.bind(this);
    this.Commit = this.Commit.bind(this);
  }
  Redirect(request: RedirectRequest): Promise<RedirectResponse> {
    const data = RedirectRequest.toJSON(request);
    const promise = this.rpc.request(this.service, "Redirect", data);
    return promise.then((data) => RedirectResponse.fromJSON(data));
  }

  Numbering(request: NumberingRequest): Promise<NumberingResponse> {
    const data = NumberingRequest.toJSON(request);
    const promise = this.rpc.request(this.service, "Numbering", data);
    return promise.then((data) => NumberingResponse.fromJSON(data));
  }

  Commit(request: CommitRequest): Promise<CommitResponse> {
    const data = CommitRequest.toJSON(request);
    const promise = this.rpc.request(this.service, "Commit", data);
    return promise.then((data) => CommitResponse.fromJSON(data));
  }
}

interface Rpc {
  request(service: string, method: string, data: unknown): Promise<unknown>;
}